Add files via upload
Using the power of HTML and CSS, I created this eye-catching profile card as a unique way to showcase my professional background and expertise.
Creating this profile card was a delightful journey that allowed me to combine my coding skills with my artistic side. It's a testament to my dedication to continuous learning and my commitment to presenting information in an engaging and visually appealing way.